gem::sky130_pdk

Function load_pdk_models

pub fn load_pdk_models(
    pdk_cells_path: &Path,
    cell_types: &[String],
) -> PdkModels

Load all PDK models needed for a set of cell types.

Arguments

pdk_cells_path - Path to the cells/ directory of the PDK

cell_types - Set of cell type names that need models (e.g. [“o21ai”, “nand2”, …])

Returns

A PdkModels containing all loaded models and UDPs.